Migrate mtoShipmentTypeToFriendlyDisplay over into utils/formatters
